Improve performance of task 3

+++ src/main/java/a3/Main.java
@@ -41,7 +41,7 @@ public class Main {
             e.printStackTrace();
         }
         
-        long binomialCoefficient = initializingThread.getSolution();
+        int binomialCoefficient = initializingThread.getSolution();
         System.out.println("Der Binomialkoeffizient ist: " + binomialCoefficient);;
 
     }
